# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fire Prevention API",
        "description": "API documentation with JWT authentication",
        "version": "1.0.0"
    },
    "paths": {
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get all alerts",
                "description": "Returns a list of all alerts with their associated device information",
                "operationId": "6ac30b403570528057748370e2ecd252",
                "responses": {
                    "200": {
                        "description": "List of alerts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No alerts found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No alerts found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Create a new alert",
                "description": "Creates a new alert with the provided information",
                "operationId": "583ead7de5a61ba68a245e8d45325f4b",
                "requestBody": {
                    "description": "Alert creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_token",
                                    "type",
                                    "value",
                                    "max_value"
                                ],
                                "properties": {
                                    "device_token": {
                                        "description": "Token of the device",
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "type": {
                                        "description": "Type of the alert",
                                        "type": "string",
                                        "enum": [
                                            "Temperatura",
                                            "Humedad",
                                            "Nivel de humo",
                                            "Nivel de gas"
                                        ],
                                        "example": "Temperatura"
                                    },
                                    "value": {
                                        "description": "Current value",
                                        "type": "number",
                                        "format": "float",
                                        "example": 45.5
                                    },
                                    "max_value": {
                                        "description": "Maximum allowed value",
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alert created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alert created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "device_token": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The device token is required"
                                            ]
                                        },
                                        "type": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The type must be one of: Temperatura, Humedad, Nivel de humo, Nivel de gas"
                                            ]
                                        },
                                        "value": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The value must be a number"
                                            ]
                                        },
                                        "max_value": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The max value must be a number"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/alerts/{id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get a single alert",
                "description": "Returns detailed information about a specific alert",
                "operationId": "e738888eb649a04431b976cdb22be3fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the alert to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alert not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Update an alert",
                "description": "Updates an existing alert with the provided information",
                "operationId": "5608d1492b83b9178f013e72e3468c48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the alert to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Alert update data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "value",
                                    "max_value"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Type of the alert",
                                        "type": "string",
                                        "enum": [
                                            "Temperatura",
                                            "Humedad",
                                            "Nivel de humo",
                                            "Nivel de gas"
                                        ],
                                        "example": "Humedad"
                                    },
                                    "value": {
                                        "description": "Current value",
                                        "type": "number",
                                        "format": "float",
                                        "example": 60
                                    },
                                    "max_value": {
                                        "description": "Maximum allowed value",
                                        "type": "number",
                                        "format": "float",
                                        "example": 70
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alert updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alert updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alert not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The type must be one of: Temperatura, Humedad, Nivel de humo, Nivel de gas"
                                            ]
                                        },
                                        "value": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The value must be a number"
                                            ]
                                        },
                                        "max_value": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The max value must be a number"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Delete an alert",
                "description": "Deletes an existing alert",
                "operationId": "acacb2da7ac666316db28483ddf92e6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the alert to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alert deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alert not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AlertResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "device_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "Temperatura",
                            "Humedad",
                            "Nivel de humo",
                            "Nivel de gas"
                        ],
                        "example": "Temperatura"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 45.5
                    },
                    "max_value": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-29T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-29T12:00:00Z"
                    },
                    "device": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Sensor Cocina"
                            },
                            "location": {
                                "type": "string",
                                "example": "Cocina Principal"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}